Extract Loaders
Loading and parsing now has enough options to be treated as a separate thing. Before all the code was merged right into the static methods of types themselves. Separate into a class hierarchy instead. Get rid of surrogate JSLite and CSSLite types and use configuration for JSLoader and CSSLoader instead.