Don't branch on naming of JSMock module IDs based on presence of package.json
Previously we would name JSMock resources differently if they existed within a
project with a package.json file. Attempting to match the name of a JSMock with
its source module (in a CommonJS environment), however, is completely broken
because modules can have many names (relative paths or project-rooted absolute
paths).

Since this was already broken, I'm changing JSMockLoader to name JSMock
resources after the filename exclusively since we actually have use cases for
that at the moment (and not for the alternative).

If the need for manual mocks inside a CJS project presents itself in the future
(it probably will), we should probably solve this problem by making a new
CommonJSMock resource type + loader to handle them separately from
filename-based manual mocks.