Fix Windows compatibility
There was 3 issues that broke node haste on windows.

`fast-path.normalize` returns a lowercase drive letter which is different
from node's `path.normalize` and caused a bunch of issues everywhere. To
fix I wrapped path in a module and return the node one on Windows. This
allows keeping the fast-path optimization on mac and linux. I'll try to
get it fixed upstream so this is no longer necessary.

`'/package.json'` string in HasteMap.js needs to be `path.sep +
'package.json'`

`'/node_modules/'` string in DependencyGraphHelpers.js needs to be
`path.sep + 'node_modules' + path.sep`

Test plan
Run the UIExplorer example in react-native on both Windows and Mac. Had to
make a few tweaks to RN so it works with node-haste 2.1.1