Make the order of sub-dependencies deterministic.
This makes the order of dependencies deterministic, if modules have a common set of subdependencies.

In the dependency tree
```
├── a
│   ├── b
│   └── c
└── d
    ├── b
    └── c
```

The order of dependencies did depend on whether `a` or `d` finished scanning their dependencies first.

It could either be `a, b, c, d` or `a, d, b, c`. This change guarantees a depth-first order, for this tree that would be `a, b, c, d`