Timeout error raised by FileWatcher is silently swallowed
If the MAX_WAIT_TIME timeout is reached, the error that is thrown by
FileWatcher is swallowed by the _loading Promise and thus it nevers gets
to the caller. This can be fixed by adding a 'done' block to this
promise.
Also, some of the messages that are displayed to the user should be a
bit more clearer.