Made all FS async operations timeoutable
In large code bases we have problems of graceful-fs read/write operations to hang infinitely.
To mitigate the problem and to be able to find out the reason I implemented 5 seconds timeout for all those fs operations.

This is quite ad-hoc, I think it would be cleaner to introduce timeouts directly to graceful-fs library but I am follwoing David Aurelio's lead here.